test(examples): show custom scan generators
